requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
seaborn
opencv-python
tqdm
pillow
